refactor: separated fragments to another packages, created FragmentFactory for the fragments
