+++ app/src/main/java/ksc/campus/tech/kakao/map/views/fragments/SearchResultFragment.kt
@@ -1,4 +1,4 @@
-package ksc.campus.tech.kakao.map.views
+package ksc.campus.tech.kakao.map.views.fragments
 
 import android.content.Context
 import android.os.Bundle
@@ -8,7 +8,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
 import androidx.core.view.isVisible
-import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
